# docs/swagger.yaml
definitions:
  domain.Task:
    properties:
      created_at:
        format: date-time
        type: string
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /tasks:
    post:
      consumes:
      - application/json
      description: Добавление Задачи
      parameters:
      - description: Данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание Таски
      tags:
      - task
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление задачи по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - task
    get:
      consumes:
      - application/json
      description: Получение задачи по её идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу по ID
      tags:
      - task
    put:
      consumes:
      - application/json
      description: Обновление задачи по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить задачу
      tags:
      - task
swagger: "2.0"